Exporting a Differ object that could be vendored in to other libraries
